--- accounts/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from . import views
-
-app_name = "accounts"
-
-urlpatterns = [
-    path("", views.IndexView.as_view(), name="index"),
-]
