+++ articles/views.py
@@ -1,3 +1,6 @@
-from django.shortcuts import render
+from django.views.generic import TemplateView
+
 
 # Create your views here.
+class IndexView(TemplateView):
+    template_name = "articles/index.html"
