+++ contents/templates/articles/index.html
@@ -6,13 +6,13 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>ほしのなか政府</title>
-  <link rel="stylesheet" href="{% sass_src 'css/accounts/index.scss' %}">
+  <link rel="stylesheet" href="{% sass_src 'css/articles/index.scss' %}">
   <link rel="icon" href="{% static 'image/favicon.ico' %}" type="image/x-icon">
 </head>
 <body>
   <header>
     <nav>
-      <a href="{% url 'accounts:index' %}"><img src="{% static 'image/logo.png' %}" height="100px"></a>
+      <a href="{% url 'articles:index' %}"><img src="{% static 'image/logo.png' %}" height="100px"></a>
     </nav>
   </header>
   <h1>ほしのなか政府 トップページ</h1>
